This idea will help others and build us a company to keep helping
Its been on the books for a while but the tech is just at the point it can all come together and become  a green sustainable brand . How do you want history to remember you? Like the ones that get forgotten over time or maybe along side , Leonardo Devin chi or tesla and Elon musk I think its doable with the right people and drive...  So if that sounds like you even if your still stuck at home and in school but in the future let us know and join us if you want. Thanks Ryan